Add tests

+++ bonus.error_handling/error_handling.cpp
@@ -19,7 +19,7 @@ TEST(ErrorHandling, ErrorHandlingTest) {
 
     auto entityIds = gis.loadMapFile("map-with-errors.json");
 
-    ASSERT_EQ(entityIds.size(), 1);
+    ASSERT_EQ((int) entityIds.size(), 1);
     ASSERT_TRUE(logContainsError("Couldn't load invalid entity - JSON entity doesn't contain name"));
     ASSERT_TRUE(logContainsError("Entity with id '1' already exists"));
     ASSERT_TRUE(logContainsError("Couldn't load invalid entity - Invalid coordinate in JSON"));
